feat(python): refactor
