+++ python/dataverse_sdk/client.py
@@ -3,7 +3,7 @@ from typing import Optional
 from .apis.backend import BackendAPI
 from .constants import DataverseHost
 from .exceptions.client import ClientConnectionError
-from .schemas.api import AttributeAPISchema, OntologyAPISchema, OntologyClassAPISchema
+from .schemas.api import AttributeAPISchema, OntologyAPISchema
 from .schemas.client import Ontology, Project, Sensor
 
 
@@ -18,7 +18,6 @@ class DataverseClient:
     ) -> None:
         """
         Instantiate a Dataverse client.
-
         Parameters
         ----------
         host : DataverseHost
@@ -26,14 +25,12 @@ class DataverseClient:
         password : Optional[str], optional
         access_token : Optional[str], optional
         refresh_token : Optional[str], optional
-
         Raises
         ------
         ValueError
         """
         if host not in DataverseHost:
             raise ValueError("Invalid dataverse host, if the host is available?")
-
         self.host = host
         self._api_client = None
         self._init_api_client(
@@ -50,6 +47,7 @@ class DataverseClient:
         access_token: Optional[str] = None,
         refresh_token: Optional[str] = None,
     ) -> None:
+
         try:
             self._api_client = BackendAPI(
                 host=self.host,
@@ -65,64 +63,52 @@ class DataverseClient:
         self, name: str, ontology: Ontology, sensors: list[Sensor]
     ) -> Project:
         """Creates project from client data
-
         Parameters
         ----------
         name : str
             name of current project
         ontology : Ontology
-            the Ontology BaseModel data from client
-        sensors : list[ClientSensor]
-            the list of Sensor BaseModel data from client
-
+            the Ontology basemodel data of current project
+        sensors : list[Sensor]
+            the list of Sensor basemodel data of current project
         Returns
         -------
-        Project
-            return Project BaseModel
+        project : Project
+            Project basemodel from host response for client usage
 
         Raises
         ------
         ClientConnectionError
             raise exception if there is any error occurs
         """
-
-        ontology_data: dict = ontology.dict(exclude_none=True)
-        classes_data_list: list[OntologyClassAPISchema] = []
-        project_data: Project = {}
-        try:
-            # remove `id` field in OntologyClass and Attribute
-            for cls_ in ontology_data["classes"]:
-                cls_.pop("id", None)
-                if not (cur_attrs := cls_.pop("attributes", None)):
-                    classes_data_list.append(cls_)
-                    continue
-                new_attribute_list: list[AttributeAPISchema] = []
-                for attr in cur_attrs:
-                    attr.pop("id", None)
-                    if attr["type"] != "option":
-                        new_attribute_list.append(attr)
-                        continue
-                    attr["option_data"] = [
-                        opt_data["value"] for opt_data in attr.pop("options", [])
-                    ]
-                    new_attribute_list.append(attr)
-                cls_["attribute_data"] = new_attribute_list
+        raw_ontology_data: dict = ontology.dict(exclude_none=True)
+        classes_data_list: list[dict] = []
+        # remove `id` field in OntologyClass and Attribute
+        for cls_ in raw_ontology_data.pop("classes", []):
+            cls_.pop("id", None)
+            if not (cur_attrs := cls_.pop("attributes", None)):
                 classes_data_list.append(cls_)
-
-            ontology_data = OntologyAPISchema(
-                **{
-                    "name": ontology_data["name"],
-                    "image_type": ontology_data["image_type"],
-                    "ontology_classes_data": classes_data_list,
-                }
-            ).dict(exclude_none=True)
-
+                continue
+            new_attribute_list: list[AttributeAPISchema] = []
+            for attr in cur_attrs:
+                attr.pop("id", None)
+                if attr["type"] != "option":
+                    new_attribute_list.append(attr)
+                    continue
+                attr["option_data"] = [
+                    opt_data["value"] for opt_data in attr.pop("options", [])
+                ]
+                new_attribute_list.append(attr)
+            cls_["attribute_data"] = new_attribute_list
+            classes_data_list.append(cls_)
+        raw_ontology_data["ontology_classes_data"] = classes_data_list
+        ontology_data = OntologyAPISchema(**raw_ontology_data).dict(exclude_none=True)
+        sensor_data = [sensor.dict(exclude_none=True) for sensor in sensors]
+        try:
             project_data: dict = self._api_client.create_project(
                 name=name,
                 ontology_data=ontology_data,
-                sensor_data=[
-                    {"name": sensor.name, "type": sensor.type} for sensor in sensors
-                ],
+                sensor_data=sensor_data,
             )
         except Exception as e:
             raise ClientConnectionError(f"Failed to create the project: {e}")
