+++ python/dataverse_sdk/schemas/client.py
@@ -1,25 +1,24 @@
-from typing import Optional
+import re
+from typing import Optional, Union
 
-from pydantic import validator
+from pydantic import BaseModel, validator
 
-from .base import (
-    AttributeType,
-    BaseAttributeOptionSchema,
-    BaseAttributeSchema,
-    BaseOntologyClassSchema,
-    BaseOntologySchema,
-    BaseProjectSchema,
-    BaseSensorSchema,
-)
+from .common import AttributeType, OntologyImageType, OntologyPcdType, SensorType
 
 
-class AttributeOption(BaseAttributeOptionSchema):
-    pass
+class AttributeOption(BaseModel):
+    id: Optional[int] = None
+    value: Union[str, float, int, bool]
 
 
-class Attribute(BaseAttributeSchema):
+class Attribute(BaseModel):
+    id: Optional[int] = None
+    name: str
     options: Optional[list[AttributeOption]] = None
     type: AttributeType
+
+    class Config:
+        use_enum_values = True
 
     @validator("type")
     def option_data_validator(cls, value, values, **kwargs):
@@ -31,18 +30,48 @@ class Attribute(BaseAttributeSchema):
         return value
 
 
-class Sensor(BaseSensorSchema):
-    pass
+class Sensor(BaseModel):
+    id: Optional[int] = None
+    name: str
+    type: SensorType
+
+    class Config:
+        use_enum_values = True
 
 
-class OntologyClass(BaseOntologyClassSchema):
-    attributes: Optional[list[BaseAttributeSchema]] = None
+class OntologyClass(BaseModel):
+    id: Optional[int] = None
+    name: str
+    color: str
+    rank: int
+    attributes: Optional[list[Attribute]] = None
+
+    @validator("color", each_item=True)
+    def color_validator(cls, value):
+        if not value.startswith("#") or not re.search(
+            r"\b[a-zA-Z0-9]{6}\b", value.lstrip("#")
+        ):
+            raise ValueError(
+                f"Color field needs starts with `#` and has 6 digits behind it, get : {value}"
+            )
+        return value
 
 
-class Ontology(BaseOntologySchema):
+class Ontology(BaseModel):
+    id: Optional[int] = None
+    name: str
+    image_type: Optional[OntologyImageType] = None
+    pcd_type: Optional[OntologyPcdType] = None
     classes: Optional[list[OntologyClass]] = None
 
+    class Config:
+        use_enum_values = True
 
-class Project(BaseProjectSchema):
+
+class Project(BaseModel):
+    id: Optional[int] = None
+    name: str
+    description: Optional[str] = None
+    ego_car: Optional[str] = None
     ontology: Ontology
     sensors: list[Sensor]
